Add test interfaces

+++ src/main/scala/converters/pixel/RGBToGrayscaleConverter.scala
@@ -4,6 +4,6 @@ import app.models.pixel.{GrayscalePixel, RGBPixel}
 
 class RGBToGrayscaleConverter extends PixelConverter[RGBPixel, GrayscalePixel] {
   override def convert(item: RGBPixel): GrayscalePixel = {
-    GrayscalePixel((0.3 * item.red) + (0.59 * item.green) + (0.11 * item.blue))
+    GrayscalePixel(((0.3 * item.red) + (0.59 * item.green) + (0.11 * item.blue)).toInt)
   }
 }
